// src/types/PacienteData.ts
export type PacienteData = {
    id: number;
    nomecompleto: string;
    datadenascimento: string;
    idade: number;
    VAD: boolean;
    alergia: boolean;
    alergiaLatex: boolean;
    prontuario: string;
    CPF: string;
    RG: string;
    telefone: string;
    observacao: string;
    createdAt: string;
    updatedAt: string;
};